Create html in build process
