+++ src/common.js
@@ -1,7 +1,2 @@
 export const EVENT_KEY = "rn-web-bridge";
 export const ROOT_ID = "root";
-
-export type Message<T> = {
-  type: string;
-  data: T;
-};
