+++ src/types.ts
@@ -1,0 +1,4 @@
+export type Message<T> = {
+  type: string;
+  data: T;
+};
